Extract common methods for CustomSAC and CustomDQN. And other changes
- Check for existing tags in wandb project
