+++ conf/combo/dqn_cartpole.yaml
@@ -20,8 +20,6 @@ agent:
     nn_width: 256
     exploration_final_eps: 0.04
 eval:
-  # seed: 456
-  # freq_eval_interval: 5000
   freq_eval_episodes: 10
   ls_eval_episodes: 20
   final_eval_episodes: 20
@@ -31,6 +29,6 @@ viz: # visualization (e.g. in wandb)
   max_return: 500
   max_ep_length: 500
   hist_bins: 50
-phases:
-  - 30000
-  - 60000
+# phases:
+#   - 30000
+#   - 60000
