+++ conf/ls/sac.yaml
@@ -1,0 +1,15 @@
+# @package _global_
+
+ls:
+  type: Sobol
+  dims:
+    - learning_rate:
+        type: Log
+        base: 1000
+        lower: 0.0001
+        upper: 0.1
+    - neg_gamma:
+        type: Log
+        base: 5
+        lower: 0.0001
+        upper: 0.2 # or 0.5
