+++ conf/phases/quick.yaml
@@ -6,6 +6,7 @@ phases:
   - 500
 agent:
   hps:
+    # WARNING these overwrite values from the combo configuration!
     learning_starts: 100 # start training quite early
     train_freq: 50 # update policy much more often
 eval:
